+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- **Vue**: Make sure only CSS files are injected to head
+
 ## 0.0.9
 
 - Env variables are now available under both `process.env` and `import.meta.env`, there're also `import.meta.env.DEV`, `import.meta.env.PROD`, `import.meta.env.BASE_URL` and `import.meta.env.MODE`, improving Vite compatibility.